+++ src/sub_commands/c_dev/routes/virtualcdn.rs
@@ -1,5 +1,7 @@
-use axum::Router;
-use crate::common::get_torytis_dir_path_buf;
+use std::fs;
+
+use axum::{body::Body, http::StatusCode, response::Response, routing::get, Router};
+use crate::{common::{get_torytis_dir_path_buf, get_working_dir_path_buf}, structs::replacer::Replacer};
 use tower_http::services::ServeDir;
 
 pub fn routes() -> Router {
@@ -8,16 +10,21 @@ pub fn routes() -> Router {
 
     Router::new()
         // .route("/script.js", get(root_route))
+        .route("/style.css", get(style_css_route))
         .nest_service("/", serve_dir)
 }
 
-// async fn root_route() -> Response {
-//     let content = get_script_js_content();
+async fn style_css_route() -> Response {
+    let style_css_path_buf = get_working_dir_path_buf().join(".torytis").join("style.css");
+    let style_css_path = style_css_path_buf.as_path();
 
-//     // let skin_html_content_str = skin_html_content.as_str();
-//     return Response::builder()
-//       .status(StatusCode::OK)
-//       .header("Content-Type", "text/css")
-//       .body(Body::from(content))
-//       .unwrap();
-// }
+    let content = fs::read_to_string(style_css_path).unwrap();
+    let replacer = Replacer::new(&content);
+    replacer.apply_images_to_virtualcdn();
+
+    return Response::builder()
+      .status(StatusCode::OK)
+      .header("Content-Type", "text/css")
+      .body(Body::from(replacer.get_html()))
+      .unwrap();
+}